+++ android/src/main/kotlin/com/greenbits/datadog_flutter/DatadogFlutterPlugin.kt
@@ -37,26 +37,6 @@ public class DatadogFlutterPlugin: FlutterPlugin, MethodCallHandler {
     channel.setMethodCallHandler(DatadogFlutterPlugin());
   }
 
-  // This static function is optional and equivalent to onAttachedToEngine. It supports the old
-  // pre-Flutter-1.12 Android projects. You are encouraged to continue supporting
-  // plugin registration via this function while apps migrate to use the new Android APIs
-  // post-flutter-1.12 via https://flutter.dev/go/android-project-migration.
-  //
-  // It is encouraged to share logic between onAttachedToEngine and registerWith to keep
-  // them functionally equivalent. Only one of onAttachedToEngine or registerWith will be called
-  // depending on the user's project. onAttachedToEngine or registerWith must both be defined
-  // in the same class.
-  companion object {
-    @JvmStatic lateinit var context : Context
-
-    @JvmStatic
-    fun registerWith(registrar: Registrar) {
-      context = registrar.activity().getApplication()
-      val channel = MethodChannel(registrar.messenger(), "plugins.greenbits.com/datadog_flutter")
-      channel.setMethodCallHandler(DatadogFlutterPlugin())
-    }
-  }
-
   override fun onMethodCall(@NonNull call: MethodCall, @NonNull result: Result) {
     when {
       call.method == "initWithClientToken" -> {