+++ CHANGELOG.md
@@ -1,9 +1,17 @@
 ## Unreleased
 
+## 2.0.0-beta.2
+
+* **BREAKING CHANGE** Remove support for Flutter 1.12 to add support Flutter 3 on Android. This only affects apps using Flutter <1.12.
+
 ## 2.0.0-beta.1
 
 * **BREAKING CHANGE** Android logs that were mapped under `Level.fine` and below are now mapped as "debug" instead of "verbose." Similarly, `Level.INFO` is now mapped to "info" instead of "debug." Log levels for Android and iOS that were `Level.SHOUT` are now mapped to "debug." While this change does not affect Flutter implementations it will affect how your logs are ingested by Datadog. (#96)
 * **BREAKING CHANGE** A new pod for iOS was introduced for Crash Reporting. In order to use this pod, make sure `DatadogSDKCrashReporting` appears in your `podfile.lock`. If it does not, remove `datadog_flutter` from your `pubspec.yaml` and run `flutter pub get`. Then readd it and run `flutter pub get; pushd ios; bundle exec pod install --repo-update; popd`.
+
+## 1.7.4
+
+* Support Flutter 3 on Android
 
 ## 1.7.3+2